% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_acs-tables.R
\docType{data}
\name{acs_educ_NY}
\alias{acs_educ_NY}
\alias{acs_race_NY}
\title{Three-way ACS Tables for New York}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1890 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4320 rows and 6 columns.
}
\source{
Kyle Walker and Matt Herman (2021). tidycensus: Load US Census Boundary and Attribute Data as
'tidyverse' and 'sf'-Ready Data Frames. R package version 0.11.4.
\url{https://CRAN.R-project.org/package=tidycensus}
}
\usage{
acs_educ_NY

acs_race_NY
}
\description{
Table of counts by CD from the 2018 one-year ACS in New York. We load the
variables that roughly partition gender, age, and race. We get another
three way table of gender, age, and education. That is, this is the
output of \code{ccesMRPprep::get_acs_cces(acscodes_age_sex_race)} and
\code{ccesMRPprep::get_acs_cces(acscodes_age_sex_educ)}.
}
\examples{
 library(tibble)
 acs_race_NY
 acs_educ_NY

}
\keyword{datasets}
