% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_helper.R
\name{formula_parts}
\alias{formula_parts}
\title{Decompose formula into parts useful for subsequent analyses}
\usage{
formula_parts(formula)
}
\arguments{
\item{formula}{A representation of the aggregate imputation or "outcome" model,
of the form \verb{X_\{K\} ~ X_1 + ... X_\{K - 1\}}}
}
\description{
Outputs a list. You can use \code{base::list2env} on the output to release
the items of the list on your environment.
}
\examples{
 synthjoint:::formula_parts(race ~ female + age + edu)
}
\keyword{internal}
