% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_helper.R
\name{predict_longer}
\alias{predict_longer}
\title{Tidy joint probabilities}
\usage{
predict_longer(
  fit,
  poptable,
  microdata,
  X_form,
  X_vars,
  area_var,
  count_var,
  outcome_var
)
}
\arguments{
\item{fit}{A model of class bmlogit/emlogit}

\item{poptable}{The population table, collapsed in terms of counts. Must contain
all variables in the RHS of \code{formula}, as well as the variables specified in
\code{area_var} and \code{count_var} below.}

\item{microdata}{The survey table that the multinomial model will be built off.
Must contain all variables in the LHS and RHS of \code{formula}.}

\item{area_var}{A character vector of the area of interest.}

\item{count_var}{A character variable that specifies which variable in \code{poptable}
indicates the count}

\item{outcome_names}{A character vector of names that correspond to each level
of the outcome of the multinomial. Must be manually set to the same ordering as
the fitted objects.}
}
\description{
Tidy joint probabilities
}
\details{
Internal function to predict from a emlogit/bmlogit object, then reshape to
the population of interest. See \code{synth_mlogit()} and \code{synth_bmlogit()}
}
\keyword{internal}
