% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_rake.R
\name{rake_spf}
\alias{rake_spf}
\title{Single Proportional Fitting (as opposed to Iterative) with microdata}
\usage{
rake_spf(outcome_var, data, fix_to, area_var, count_var = "count")
}
\arguments{
\item{outcome_var}{A string for the variable to match the proportion to}

\item{data}{Data of microdata or tables whose margin on \code{outcom_var} must be fixed}

\item{fix_to}{A dataset with only marginal counts or proportions of the outcome
in question, by each area. Proportions will be corrected so that the margins
of the synthetic joint will match these, with a simple ratio.}

\item{area_var}{A character vector of the area of interest.}

\item{count_var}{A character variable that specifies which variable in \code{poptable}
indicates the count}
}
\description{
With microdata or a table that includes both
}
