% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_rake.R
\name{rake_target}
\alias{rake_target}
\title{Create long table for rake weighitng}
\usage{
rake_target(formula, poptable, newtable, area_var, count_var = "count")
}
\arguments{
\item{formula}{A representation of the aggregate imputation or "outcome" model,
of the form \verb{X_\{K\} ~ X_1 + ... X_\{K - 1\}}}

\item{poptable}{The population table, collapsed in terms of counts. Must contain
all variables in the RHS of \code{formula}, as well as the variables specified in
\code{area_var} and \code{count_var} below.}

\item{newtable}{A dataset that contains marginal counts or proportions. Will be
collapsed internally to get simple proportions.}

\item{area_var}{A character vector of the area of interest.}

\item{count_var}{A character variable that specifies which variable in \code{poptable}
indicates the count}
}
\value{
A long table that, for each districts, list the target distribution
of the outcome (\code{variable = "outcome"}) and the cell of the known Xs (\verb{variable = "Xs}).
This is consistent with the dataframe format for targets in the
\code{autumn} package \url{https://github.com/aaronrudkin/autumn}
}
\description{
Create long table for rake weighitng
}
\examples{
library(dplyr)
library(ccesMRPprep)

# suppose we want know the distribution of (age x female) and we know the
# distribution of (race), by CD, but we don't know the joint of the two.

race_target <- count(acs_race_NY, cd, race, wt = count, name = "count")

rake_target(race ~ age + female,
            poptable = acs_race_NY,
            newtable = race_target,
            area_var = "cd")

}
