% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_bmlogit.R
\name{synth_bmlogit}
\alias{synth_bmlogit}
\title{Synthetic joint estimation with balancing constraint}
\source{
Soichiro Yamauchi and Shiro Kuriwaki (2021). bmlogit: Multinomial logit with
balancing constraints. R package version 0.0.3.
}
\usage{
synth_bmlogit(
  formula,
  microdata,
  poptable,
  fix_to,
  fix_by_area = any(area_var \%in\% colnames(fix_to)),
  area_var,
  count_var = "count",
  tol = 0.05
)
}
\arguments{
\item{formula}{A representation of the aggregate imputation or "outcome" model,
of the form \verb{X_\{K\} ~ X_1 + ... X_\{K - 1\}}}

\item{microdata}{The survey table that the multinomial model will be built off.
Must contain all variables in the LHS and RHS of \code{formula}.}

\item{poptable}{The population table, collapsed in terms of counts. Must contain
all variables in the RHS of \code{formula}, as well as the variables specified in
\code{area_var} and \code{count_var} below.}

\item{fix_to}{A dataset with only marginal counts or proportions of the outcome
in question, by each area. Proportions will be corrected so that the margins
of the synthetic joint will match these, with a simple ratio.}

\item{fix_by_area}{logical, whether to fix to targets area by area. Defaults to
TRUE if \code{area_var} is a variable in \code{fix_to}. If \code{FALSE}, collapses the
input to a single target.}

\item{area_var}{A character vector of the area of interest.}

\item{count_var}{A character variable that specifies which variable in \code{poptable}
indicates the count}

\item{tol}{Tolerance for balance}
}
\description{
Imputes cells with a balancing constraint, using Yamauchi's algorithm.
}
\examples{

library(dplyr)
library(ccesMRPprep)

# can take a few minutes if fix_by_area = TRUE (the default)
educ_target <- count(acs_educ_NY, cd, educ, wt = count, name = "count")

educ_target
acs_race_NY

pop_syn <- synth_bmlogit(educ ~ race + age + female,
                         microdata = cc18_NY,
                         fix_to = educ_target,
                         poptable = acs_race_NY,
                         area_var = "cd")
pop_syn

}
\seealso{
\code{synth_mlogit()}
}
